smaller spec files
