+++ src/SDKs/JavaScript/test/spec/common.spec.js
@@ -1,0 +1,6 @@
+const exampleConfig = { 
+      "netherClientId" : "Kristofer",
+      "netherClientSecret" : "Anko",
+      "facebookAppId" : "Stuart",
+      "netherBaseUrl" : "https://andy.and.james/from/elastacloud"
+    };
